Migrate entity ID types from Integer to Long
- Changed IDs in Book and Genre entities from Integer to Long
- Updated related DTOs and repository interfaces
- Added Flyway migrations:
  * V3__change_book_id_type.sql
  * V4__change_genre_id_type.sql

+++ src/main/java/shop/dto/book/CreateEditBookDto.java
@@ -41,7 +41,7 @@ public class CreateEditBookDto {
 	
 	private Set<Long> authors = new HashSet<>();
 	
-	private Set<Integer> genres = new HashSet<>();
+	private Set<Long> genres = new HashSet<>();
 	
 	@Override
 	public String toString() {
